# public/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Holodeck Router UI API",
    "description": "REST API for managing VCF 9 Holodeck deployments, Day 2 operations, reservations, commands, user management, and email notifications.",
    "version": "1.2.0"
  },
  "servers": [
    { "url": "/", "description": "Current instance" }
  ],
...
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "holodeck-token",
        "description": "JWT token set as httpOnly cookie after login"
      }
    },
...
  },
  "security": [{ "cookieAuth": [] }],
  "paths": {
...
    "/api/depot": {
      "get": {
        "tags": ["Depot"],
        "summary": "Scan depot or check status",
        "description": "action=check: combined scan + Broadcom list in a single SSH call (pass version param to filter). action=status: check download tmux session. No action: legacy file listing.",
        "parameters": [
          { "name": "action", "in": "query", "schema": { "type": "string", "enum": ["check", "status"] }, "description": "check=combined scan+list, status=download session status" },
          { "name": "version", "in": "query", "schema": { "type": "string" }, "description": "VCF version to filter components (for action=check)" },
          { "name": "session", "in": "query", "schema": { "type": "string" }, "description": "tmux session name (for action=status)" }
        ],
        "responses": {
          "200": {
            "description": "Depot scan results or session status",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "components": { "type": "object", "additionalProperties": true, "description": "Component name -> {files, totalSize}" },
                    "hasMetadata": { "type": "boolean" },
                    "hasVsanHcl": { "type": "boolean" },
                    "totalSize": { "type": "string" },
                    "listOutput": { "type": "string", "description": "Raw Broadcom list output" },
                    "listSuccess": { "type": "boolean", "description": "Whether Broadcom list succeeded" },
                    "alive": { "type": "boolean" },
                    "output": { "type": "string" }
                  }
                }
              }
            }
          },
          "403": { "description": "Forbidden — requires labadmin or superadmin role" }
        }
      },
      "post": {
        "tags": ["Depot"],
        "summary": "Depot operations",
        "description": "download=start download, kill=cancel download, cleanup-list=find files for a version, cleanup-delete=delete specific files.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["action"],
                "properties": {
                  "action": { "type": "string", "enum": ["download", "kill", "cleanup-list", "cleanup-delete"], "description": "Operation to perform" },
                  "vcfVersion": { "type": "string", "description": "VCF version for download" },
                  "session": { "type": "string", "description": "tmux session name for kill" },
                  "cleanupVersion": { "type": "string", "description": "Version to find/delete files for (cleanup actions)" },
                  "files": { "type": "array", "items": { "type": "string" }, "description": "File paths to delete (cleanup-delete)" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Operation result",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "output": { "type": "string" },
                    "session": { "type": "string" },
                    "message": { "type": "string" },
                    "files": { "type": "array", "items": { "type": "string" }, "description": "Matching files (cleanup-list)" },
                    "count": { "type": "integer" },
                    "deleted": { "type": "integer" }
                  }
                }
              }
            }
          },
          "400": { "description": "Missing parameters" },
          "403": { "description": "Forbidden — requires labadmin or superadmin role" },
          "409": { "description": "Download already in progress" }
        }
      }
    }
  }
}